+++ examples/flutter/lib/main.dart
@@ -3,9 +3,12 @@ import 'package:flutter/material.dart';
 import 'router.dart';
 
 void main() {
-  runApp(GeneratedApp(
+  runApp(MaterialApp.router(
+    debugShowCheckedModeBanner: false,
     themeMode: ThemeMode.system,
     theme: ThemeData.light(),
     darkTheme: ThemeData.dark(),
+    routerDelegate: router.routerDelegate,
+    routeInformationParser: router.routeInformationParser,
   ));
 }