+++ examples/flutter/lib/routes/about/:id.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class AccountPage extends StatelessWidget {
+  const AccountPage({
+    Key? key,
+    required this.id,
+  }) : super(key: key);
+
+  final String id;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('ID: $id'),
+    );
+  }
+}