+++ examples/flutter/lib/routes/about/guest.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class GuestPage extends StatelessWidget {
+  const GuestPage({
+    Key? key,
+    this.id,
+  }) : super(key: key);
+
+  final String? id;
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Guest'),
+    );
+  }
+}